+++ helper.py
@@ -5,31 +5,31 @@ from os import path as op
 
 
 # Изменение (filename, URL) вставкой '.mini' и изменение расширения на jpg
-def _add_mini(s):
+from PIL import Image
+
+
+def _add_mini(s, postfix=''):
     parts = s.split(".")
-    parts.insert(-1, "mini")
+    parts.insert(-1, postfix)
     if parts[-1].lower() not in ['jpeg', 'jpg', 'png']:
         parts[-1] = 'jpg'
     return ".".join(parts)
 
 
 # Удаление миниатюры с физического носителя.
-def _del_mini(p):
-    mini_path = _add_mini(p)
-    if os.path.exists(mini_path):
-        os.remove(mini_path)
+def _del_mini(p, postfix=''):
+    mini_path = _add_mini(p, postfix)
+    if 'default' not in mini_path:
+        if os.path.exists(mini_path):
+            os.remove(mini_path)
 
 
-def upload_to(instance, filename, prefix=None, unique=False, postfix=None):
+def upload_to(instance, filename, prefix=None, unique=False):
     ext = op.splitext(filename)[1]
     name = str(instance.pk or '') + filename + (str(time()) if unique else '')
-
-    basedir = 'media'
-    if prefix:
-        basedir = op.join(basedir, prefix)
     filename = md5(name.encode('utf8')).hexdigest() + ext
 
-    return op.join(basedir, prefix, filename[:2], filename[2:4], filename, postfix)
+    return op.join(prefix, filename[:2], filename[2:4], filename)
 
 
 def parse_from_error_to_json(request, form):
@@ -45,3 +45,30 @@ def parse_from_error_to_json(request, form):
         data.append(text)
     return data
 
+
+# Создание миниатюры (на основе ориг.)
+def create_mini_image(img):
+    img.thumbnail(
+        (128, 128),
+        Image.ANTIALIAS
+    )
+    return img
+
+
+# Создание форматированного изображения (на основе ориг.)
+def create_medium_image(img):
+    height = img.height
+    width = img.width
+
+    max_width = 1440
+    max_height = 960
+    if width >= max_width or height >= max_height:
+        correlation = width / height
+        height = width / correlation
+        width = max_width
+
+        img.thumbnail(
+            (width, height),
+            Image.ANTIALIAS
+        )
+    return img
